requirements.txt
sentence-transformers==2.2.0
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .

#pip install -r requirements.txt